// src/types/EmpresaAttributes.ts
export interface EmpresaAttributes {
    id: number;
    nome: string;
    nomeFantasia: string;
    cnpj: string;
    setor: string;
    enderecoSede: string;
    nomeResponsavel: string;
    contatoResponsavel: string;
    emailResponsavel: string;
    observacoes: string;
    tipo: string;
    idEmpresaSede: number;
  }
  